+++ Sources/TokamakDemo/main.swift
@@ -12,25 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// import JavaScriptKit
-// import TokamakDOM
-//
-// let document = JSObjectRef.global.document.object!
-// let body = document.body.object!
-// body.style = "margin: 0;"
-//
-// let div = document.createElement!("div").object!
-// let renderer = DOMRenderer(TokamakDemoView(), div)
-//
-// _ = body.appendChild!(div)
-
 #if os(WASI)
 import TokamakDOM
 #else
 import SwiftUI
 #endif
 
-@available(OSX 10.16, *)
+@available(OSX 10.16, iOS 14.0, *)
 struct TokamakDemoApp: App {
   @Environment(\.scenePhase) private var scenePhase
 
@@ -44,6 +32,6 @@ struct TokamakDemoApp: App {
 
 // If @main was supported for executable Swift Packages,
 // this would match SwiftUI 100%
-if #available(OSX 10.16, *) {
+if #available(OSX 10.16, iOS 14.0, *) {
   TokamakDemoApp.main()
 }
